+++ src/stock-trading-vue/static/components/sidebar/SidebarMenu.vue
@@ -32,7 +32,7 @@ const user = useUserStore()
                     </a>
                 </li>
                 <li class="nav-item">
-                    <a class="nav-link d-flex align-items-center gap-2" href="#">
+                    <a class="nav-link d-flex align-items-center gap-2" href="/trades">
                         <i class="bi bi-file-earmark-lock"></i>
                         Trade History
                     </a>